chore(apply-patch) move invocation tests
## Summary:
This PR is a pure copy and paste of tests from lib.rs into invocation.rs, to colocate logic and tests.

## Testing
- [x] Purely a test refactor
